Revert "Merge branch 'feature/foodsubscription-foodOrderApplicable' into interest_dashboard_hybrid-menu"
This reverts commit 6b95acbc3fcc7132248e8d54f32441f27be285e7, reversing
changes made to 2bb1ea62ce0090f0b01253979007209e46bb036b.

+++ pojo-wanda/src/main/java/com/stanzaliving/wanda/dtos/ChangeContractDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Getter
 public class ChangeContractDto {
 
-	private long contractId;
+	private int contractId;
 	
 	private String contractMonths;
